requirements.txt
# LangChain Core
langchain
langchain-google-genai
langchain_community
# PDF Parsing
pypdf

# Vector Database
chromadb
# Environment Variables
python-dotenv

# UI
streamlit

## Text Extraction
pdfminer.six

## For Similarity 
scikit-learn